Reorganize samplers package
