+++ clusview/samplers/parameters/polynomial_sampler.py
@@ -1,37 +1,8 @@
-from abc import ABC, abstractmethod
 from typing import Iterable
 
 import numpy as np
 
-
-class BaseSampler(ABC):
-    """
-    Base class for all parameter samplers.
-    """
-
-    def __init__(self, parameter_name: str, lower_bound: int, upper_bound: int):
-        self.parameter_name = parameter_name
-        self.lower_bound = lower_bound
-        self.upper_bound = upper_bound
-
-    @abstractmethod
-    def sample_range(self) -> Iterable:
-        pass
-
-
-class LinearSampler(BaseSampler):
-    """
-    Linear parameter sampler.
-    """
-
-    def __init__(
-        self, parameter_name: str, lower_bound: int, upper_bound: int, step: int
-    ):
-        super().__init__(parameter_name, lower_bound, upper_bound)
-        self.step = step
-
-    def sample_range(self) -> Iterable:
-        return range(self.lower_bound, self.upper_bound + 1, self.step)
+from clusview.samplers.parameters.base_parameter_sampler import BaseSampler
 
 
 class PolynomialSampler(BaseSampler):
